+++ src/BigDecimal.php
@@ -45,15 +45,11 @@ final class BigDecimal extends BigNumber
     }
 
     /**
-     * Creates a BigDecimal of the given value.
-     *
-     * @throws MathException If the value cannot be converted to a BigDecimal.
-     *
      * @psalm-pure
      */
-    public static function of(BigNumber|int|float|string $value) : BigDecimal
+    protected static function from(BigNumber $number): static
     {
-        return parent::of($value)->toBigDecimal();
+        return $number->toBigDecimal();
     }
 
     /**